+++ Examples/GLImage Demo/GLViewExample/GLViewMipmapController.m
@@ -21,8 +21,8 @@
     [super viewDidLoad];
     
     //load image and set default scale
-    _mipmapImageView.image = [GLImage imageWithContentsOfFile:@"mipmapped-image.pvr"];
-    _mipmapImageView.imageTransform = CATransform3DMakeScale(0.1f, 0.1f, 0.1f);
+    self.mipmapImageView.image = [GLImage imageWithContentsOfFile:@"mipmapped-image.pvr"];
+    self.mipmapImageView.imageTransform = CATransform3DMakeScale(0.1f, 0.1f, 0.1f);
     
     //pinch gesture
     UIPinchGestureRecognizer *pinch = [[UIPinchGestureRecognizer alloc] initWithTarget:self action:@selector(pinch:)];
@@ -32,17 +32,10 @@
 
 - (void)pinch:(UIPinchGestureRecognizer *)pinch
 {
-    CATransform3D transform = _mipmapImageView.imageTransform;
+    CATransform3D transform = self.mipmapImageView.imageTransform;
     transform = CATransform3DScale(transform, pinch.scale, pinch.scale, pinch.scale);
-    _mipmapImageView.imageTransform = transform;
+    self.mipmapImageView.imageTransform = transform;
     pinch.scale = 1.0f;
-}
-
-- (void)viewDidUnload
-{
-    [super viewDidUnload];
-    
-    self.mipmapImageView = nil;
 }
 
 @end