+++ Examples/GLImage Demo/GZIP/GZIP.h
@@ -1,7 +1,7 @@
 //
-//  NSData+GZIP.h
+//  GZIP.h
 //
-//  Version 1.0
+//  Version 1.0.2
 //
 //  Created by Nick Lockwood on 03/06/2012.
 //  Copyright (C) 2012 Charcoal Design
@@ -9,7 +9,7 @@
 //  Distributed under the permissive zlib License
 //  Get the latest version from here:
 //
-//  https://github.com/nicklockwood/Gzip
+//  https://github.com/nicklockwood/GZIP
 //
 //  This software is provided 'as-is', without any express or implied
 //  warranty.  In no event will the authors be held liable for any damages