+++ Examples/GLImage Demo/GZIP/GZIP.m
@@ -1,7 +1,7 @@
 //
-//  NSData+GZIP.m
+//  GZIP.m
 //
-//  Version 1.0
+//  Version 1.0.2
 //
 //  Created by Nick Lockwood on 03/06/2012.
 //  Copyright (C) 2012 Charcoal Design
@@ -9,7 +9,7 @@
 //  Distributed under the permissive zlib License
 //  Get the latest version from here:
 //
-//  https://github.com/nicklockwood/Gzip
+//  https://github.com/nicklockwood/GZIP
 //
 //  This software is provided 'as-is', without any express or implied
 //  warranty.  In no event will the authors be held liable for any damages
@@ -31,11 +31,11 @@
 //
 
 
-#import "NSData+GZIP.h"
+#import "GZIP.h"
 #import <zlib.h>
 
 
-#define CHUNK_SIZE 16384
+static const NSUInteger ChunkSize = 16384;
 
 
 @implementation NSData (GZIP)
@@ -53,18 +53,18 @@
         stream.total_out = 0;
         stream.avail_out = 0;
         
-        int compression = (level < 0.0f)? Z_DEFAULT_COMPRESSION: (int)roundf(level * 9);
+        int compression = (level < 0.0f)? Z_DEFAULT_COMPRESSION: (int)(roundf(level * 9));
         if (deflateInit2(&stream, compression, Z_DEFLATED, 31, 8, Z_DEFAULT_STRATEGY) == Z_OK)
         {
-            NSMutableData *data = [NSMutableData dataWithLength:CHUNK_SIZE];
+            NSMutableData *data = [NSMutableData dataWithLength:ChunkSize];
             while (stream.avail_out == 0)
             {
                 if (stream.total_out >= [data length])
                 {
-                    data.length += CHUNK_SIZE;
+                    data.length += ChunkSize;
                 }
-                stream.next_out = [data mutableBytes] + stream.total_out;
-                stream.avail_out = (uint)([data length] - stream.total_out);
+                stream.next_out = (uint8_t *)[data mutableBytes] + stream.total_out;
+                stream.avail_out = (uInt)([data length] - stream.total_out);
                 deflate(&stream, Z_FINISH);
             }
             deflateEnd(&stream);
@@ -92,7 +92,7 @@
         stream.total_out = 0;
         stream.avail_out = 0;
         
-        NSMutableData *data = [NSMutableData dataWithLength: [self length] * 1.5];
+        NSMutableData *data = [NSMutableData dataWithLength:(NSUInteger)([self length] * 1.5)];
         if (inflateInit2(&stream, 47) == Z_OK)
         {
             int status = Z_OK;
@@ -102,8 +102,8 @@
                 {
                     data.length += [self length] * 0.5;
                 }
-                stream.next_out = [data mutableBytes] + stream.total_out;
-                stream.avail_out = (uint)([data length] - stream.total_out);
+                stream.next_out = (uint8_t *)[data mutableBytes] + stream.total_out;
+                stream.avail_out = (uInt)([data length] - stream.total_out);
                 status = inflate (&stream, Z_SYNC_FLUSH);
             }
             if (inflateEnd(&stream) == Z_OK)