+++ Examples/GLModel Demo/GLModelExample/ViewController.m
@@ -11,9 +11,6 @@
 
 @implementation ViewController
 
-@synthesize navBar;
-@synthesize modelView;
-
 - (void)viewDidUnload
 {
     [super viewDidUnload];
@@ -28,106 +25,106 @@
         case 0:
         {
             //set title
-            navBar.topItem.title = @"demon.model";
+            self.navBar.topItem.title = @"demon.model";
             
             //set model
-            modelView.texture = [GLImage imageNamed:@"demon.png"];
-            modelView.blendColor = nil;
-            modelView.model = [GLModel modelWithContentsOfFile:@"demon.model"];
+            self.modelView.texture = [GLImage imageNamed:@"demon.png"];
+            self.modelView.blendColor = nil;
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"demon.model"];
             
             //set default transform
             CATransform3D transform = CATransform3DMakeTranslation(0.0f, 0.0f, -2.0f);
             transform = CATransform3DScale(transform, 0.01f, 0.01f, 0.01f);
-            transform = CATransform3DRotate(transform, -M_PI_2, 1.0f, 0.0f, 0.0f);
-            modelView.modelTransform = transform;
+            transform = CATransform3DRotate(transform, (CGFloat)-M_PI_2, 1.0f, 0.0f, 0.0f);
+            self.modelView.modelTransform = transform;
             
             break;
         }
         case 1:
         {
             //set title
-            navBar.topItem.title = @"quad";
+            self.navBar.topItem.title = @"quad";
             
             //set model
-            modelView.texture = nil;
-            modelView.blendColor = [UIColor redColor];
-            modelView.model = [GLModel modelWithContentsOfFile:@"quad.obj"];
+            self.modelView.texture = nil;
+            self.modelView.blendColor = [UIColor redColor];
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"quad.obj"];
             
             //set default transform
-            modelView.modelTransform = CATransform3DMakeTranslation(0.0f, 0.0f, -2.0f);
+            self.modelView.modelTransform = CATransform3DMakeTranslation(0.0f, 0.0f, -2.0f);
             
             break;
         }
         case 2:
         {
             //set title
-            navBar.topItem.title = @"chair.obj";
+            self.navBar.topItem.title = @"chair.obj";
             
             //set model
-            modelView.texture = [GLImage imageNamed:@"chair.tga"];
-            modelView.blendColor = nil;
-            modelView.model = [GLModel modelWithContentsOfFile:@"chair.obj"];
+            self.modelView.texture = [GLImage imageNamed:@"chair.tga"];
+            self.modelView.blendColor = nil;
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"chair.obj"];
             
             //set default transform
             CATransform3D transform = CATransform3DMakeTranslation(0.0f, 0.0f, -2.0f);
             transform = CATransform3DScale(transform, 0.01f, 0.01f, 0.01f);
             transform = CATransform3DRotate(transform, 0.2f, 1.0f, 0.0f, 0.0f);
-            modelView.modelTransform = transform;
+            self.modelView.modelTransform = transform;
             
             break;
         }
         case 3:
         {
             //set title
-            navBar.topItem.title = @"diamond.obj";
+            self.navBar.topItem.title = @"diamond.obj";
             
             //set model
-            modelView.texture = nil;
-            modelView.blendColor = [UIColor greenColor];
-            modelView.model = [GLModel modelWithContentsOfFile:@"diamond.obj"];
+            self.modelView.texture = nil;
+            self.modelView.blendColor = [UIColor greenColor];
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"diamond.obj"];
             
             //set default transform
             CATransform3D transform = CATransform3DMakeTranslation(0.0f, 0.0f, -1.0f);
             transform = CATransform3DScale(transform, 0.01f, 0.01f, 0.01f);
-            transform = CATransform3DRotate(transform, M_PI_2, 1.0f, 0.0f, 0.0f);
-            modelView.modelTransform = transform;
+            transform = CATransform3DRotate(transform, (CGFloat)M_PI_2, 1.0f, 0.0f, 0.0f);
+            self.modelView.modelTransform = transform;
             
             break;
         }
         case 4:
         {
             //set title
-            navBar.topItem.title = @"cube.obj";
+            self.navBar.topItem.title = @"cube.obj";
             
             //set model
-            modelView.texture = nil;
-            modelView.blendColor = [UIColor whiteColor];
-            modelView.model = [GLModel modelWithContentsOfFile:@"cube.obj"];
+            self.modelView.texture = nil;
+            self.modelView.blendColor = [UIColor whiteColor];
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"cube.obj"];
             
             //set default transform
             CATransform3D transform = CATransform3DMakeTranslation(0.0f, 0.0f, -1.0f);
-            transform = CATransform3DRotate(transform, M_PI_4, 1.0f, 1.0f, 0.0f);
-            modelView.modelTransform = transform;
+            transform = CATransform3DRotate(transform, (CGFloat)M_PI_4, 1.0f, 1.0f, 0.0f);
+            self.modelView.modelTransform = transform;
             
             break;
         }
         case 5:
         {
             //set title
-            navBar.topItem.title = @"ship.obj";
+            self.navBar.topItem.title = @"ship.obj";
             
             //set model
-            modelView.texture = nil;
-            modelView.blendColor = [UIColor grayColor];
-            modelView.model = [GLModel modelWithContentsOfFile:@"ship.obj"];
+            self.modelView.texture = nil;
+            self.modelView.blendColor = [UIColor grayColor];
+            self.modelView.model = [GLModel modelWithContentsOfFile:@"ship.obj"];
             
             //set default transform
             CATransform3D transform = CATransform3DMakeTranslation(0.0f, 0.0f, -15.0f);
-            transform = CATransform3DRotate(transform, M_PI + 0.4f, 0.0f, 0.0f, 1.0f);
-            transform = CATransform3DRotate(transform, M_PI_4, 1.0f, 0.0f, 0.0f);
+            transform = CATransform3DRotate(transform, (CGFloat)M_PI + 0.4f, 0.0f, 0.0f, 1.0f);
+            transform = CATransform3DRotate(transform, (CGFloat)M_PI_4, 1.0f, 0.0f, 0.0f);
             transform = CATransform3DRotate(transform, -0.4f, 0.0f, 1.0f, 0.0f);
             transform = CATransform3DScale(transform, 3.0f, 3.0f, 3.0f);
-            modelView.modelTransform = transform;
+            self.modelView.modelTransform = transform;
             
             break;
         }
@@ -143,7 +140,7 @@
                         otherButtonTitles:@"Demon", @"Quad", @"Chair", @"Diamond", @"Cube", @"Ship", nil] showInView:self.view];
 }
 
-- (void)actionSheet:(UIActionSheet *)actionSheet didDismissWithButtonIndex:(NSInteger)buttonIndex
+- (void)actionSheet:(__unused UIActionSheet *)actionSheet didDismissWithButtonIndex:(NSInteger)buttonIndex
 {
     if (buttonIndex >= 0)
     {