+++ GLView/Images/GLImageMap.h
@@ -2,7 +2,7 @@
 //  GLImageMap.h
 //
 //  GLView Project
-//  Version 1.6 beta
+//  Version 1.6
 //
 //  Created by Nick Lockwood on 04/06/2012.
 //  Copyright 2011 Charcoal Design
@@ -31,20 +31,25 @@
 //  3. This notice may not be removed or altered from any source distribution.
 //
 
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wobjc-missing-property-synthesis"
+
+
 #import "GLImage.h"
 #import "GLUtils.h"
 
 
 @interface GLImageMap : NSObject <NSFastEnumeration>
 
-+ (GLImageMap *)imageMapWithContentsOfFile:(NSString *)nameOrPath;
-+ (GLImageMap *)imageMapWithImage:(GLImage *)image data:(NSData *)data;
++ (instancetype)imageMapWithContentsOfFile:(NSString *)nameOrPath;
++ (instancetype)imageMapWithImage:(GLImage *)image data:(NSData *)data;
 
-- (GLImageMap *)initWithContentsOfFile:(NSString *)nameOrPath;
-- (GLImageMap *)initWithImage:(GLImage *)image data:(NSData *)data;
+- (instancetype)initWithContentsOfFile:(NSString *)nameOrPath;
+- (instancetype)initWithImage:(GLImage *)image data:(NSData *)data;
 
-- (NSInteger)imageCount;
-- (NSString *)imageNameAtIndex:(NSInteger)index;
+- (NSUInteger)imageCount;
+- (NSString *)imageNameAtIndex:(NSUInteger)index;
 - (GLImage *)imageAtIndex:(NSInteger)index;
 - (GLImage *)imageNamed:(NSString *)name;
 
@@ -52,3 +57,6 @@
 - (GLImage *)objectForKeyedSubscript:(NSString *)name;
 
 @end
+
+
+#pragma GCC diagnostic pop