+++ GLView/Models/GLModelView.h
@@ -2,7 +2,7 @@
 //  GLModelView.h
 //
 //  GLView Project
-//  Version 1.6 beta
+//  Version 1.6
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design
@@ -31,6 +31,11 @@
 //  3. This notice may not be removed or altered from any source distribution.
 //
 
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wobjc-missing-property-synthesis"
+
+
 #import "GLView.h"
 #import "GLLight.h"
 #import "GLModel.h"
@@ -47,3 +52,6 @@
 @property (nonatomic, copy) NSArray *lights;
 
 @end
+
+
+#pragma GCC diagnostic pop